Move data stuff aroudn
