+++ common/common_system.ts
@@ -4,12 +4,12 @@ import { SilverBulletHooks } from "./manifest.ts";
 import { buildQueryFunctions } from "./query_functions.ts";
 import { ScriptEnvironment } from "./space_script.ts";
 import { EventHook } from "../lib/plugos/hooks/event.ts";
-import { DataStore } from "$lib/datastore.ts";
-import { System } from "../lib/plugos/system.ts";
+import { DataStore } from "$lib/data/datastore.ts";
+import { System } from "$lib/plugos/system.ts";
 import { CodeWidgetHook } from "../web/hooks/code_widget.ts";
 import { PanelWidgetHook } from "../web/hooks/panel_widget.ts";
 import { SlashCommandHook } from "../web/hooks/slash_command.ts";
-import { DataStoreMQ } from "$lib/mq.datastore.ts";
+import { DataStoreMQ } from "$lib/data/mq.datastore.ts";
 
 export abstract class CommonSystem {
   system!: System<SilverBulletHooks>;
