+++ common/spaces/chunked_datastore_space_primitives.ts
@@ -1,8 +1,8 @@
 import type { SpacePrimitives } from "./space_primitives.ts";
 import { KvKey } from "$type/types.ts";
-import { KvPrimitives } from "$lib/kv_primitives.ts";
+import { KvPrimitives } from "$lib/data/kv_primitives.ts";
 import { KvMetaSpacePrimitives } from "./kv_meta_space_primitives.ts";
-import { PrefixedKvPrimitives } from "$lib/prefixed_kv_primitives.ts";
+import { PrefixedKvPrimitives } from "$lib/data/prefixed_kv_primitives.ts";
 
 /**
  * A space primitives implementation that stores files in chunks in a KV store.
