+++ common/spaces/datastore_space_primitives.test.ts
@@ -1,6 +1,6 @@
 import "https://esm.sh/fake-indexeddb@4.0.2/auto";
-import { DataStore } from "$lib/datastore.ts";
-import { IndexedDBKvPrimitives } from "$lib/indexeddb_kv_primitives.ts";
+import { DataStore } from "$lib/data/datastore.ts";
+import { IndexedDBKvPrimitives } from "$lib/data/indexeddb_kv_primitives.ts";
 import { DataStoreSpacePrimitives } from "./datastore_space_primitives.ts";
 import { testSpacePrimitives } from "./space_primitives.test.ts";
