+++ common/spaces/datastore_space_primitives.ts
@@ -1,7 +1,7 @@
 import type { SpacePrimitives } from "./space_primitives.ts";
 import { mime } from "mimetypes";
 import { FileMeta } from "$type/types.ts";
-import { DataStore } from "$lib/datastore.ts";
+import { DataStore } from "$lib/data/datastore.ts";
 
 export type FileContent = {
   name: string;
