+++ common/spaces/deno_kv_space_primitives.test.ts
@@ -1,6 +1,6 @@
-import { DenoKvPrimitives } from "$lib/deno_kv_primitives.ts";
-import { ChunkedKvStoreSpacePrimitives } from "./chunked_datastore_space_primitives.ts";
-import { testSpacePrimitives } from "./space_primitives.test.ts";
+import { DenoKvPrimitives } from "$lib/data/deno_kv_primitives.ts";
+import { ChunkedKvStoreSpacePrimitives } from "$common/spaces/chunked_datastore_space_primitives.ts";
+import { testSpacePrimitives } from "$common/spaces/space_primitives.test.ts";
 
 Deno.test("deno kv test", async () => {
   const tempFile = await Deno.makeTempFile({ suffix: ".db" });
