+++ common/spaces/encrypted_space_primitives.test.ts
@@ -1,4 +1,4 @@
-import { MemoryKvPrimitives } from "$lib/memory_kv_primitives.ts";
+import { MemoryKvPrimitives } from "$lib/data/memory_kv_primitives.ts";
 import { assert, assertEquals } from "$lib/test_deps.ts";
 import { ChunkedKvStoreSpacePrimitives } from "./chunked_datastore_space_primitives.ts";
 import { EncryptedSpacePrimitives } from "./encrypted_space_primitives.ts";
