+++ lib/data/mq.datastore.test.ts
@@ -3,7 +3,7 @@ import { assertEquals } from "$lib/test_deps.ts";
 import { DenoKvPrimitives } from "./deno_kv_primitives.ts";
 import { DataStore } from "./datastore.ts";
 import { PrefixedKvPrimitives } from "./prefixed_kv_primitives.ts";
-import { sleep } from "./async.ts";
+import { sleep } from "$lib/async.ts";
 
 Deno.test("DataStore MQ", async () => {
   const tmpFile = await Deno.makeTempFile();
