+++ lib/plugos/hooks/mq.ts
@@ -1,7 +1,7 @@
 import { Hook, Manifest } from "../types.ts";
 import { System } from "../system.ts";
 import { MQMessage } from "$type/types.ts";
-import { MessageQueue } from "$lib/mq.ts";
+import { MessageQueue } from "$lib/data/mq.ts";
 import { throttle } from "$lib/async.ts";
 
 type MQSubscription = {
