+++ lib/plugos/syscalls/datastore.ts
@@ -1,5 +1,5 @@
 import { KV, KvKey, KvQuery } from "$type/types.ts";
-import type { DataStore } from "$lib/datastore.ts";
+import type { DataStore } from "$lib/data/datastore.ts";
 import type { SysCallMapping } from "../system.ts";
 
 /**
