+++ server/http_server.ts
@@ -11,8 +11,8 @@ import { AssetBundle } from "../lib/asset_bundle/bundle.ts";
 import { FileMeta } from "$type/types.ts";
 import { ShellRequest } from "./rpc.ts";
 import { SpaceServer, SpaceServerConfig } from "./instance.ts";
-import { KvPrimitives } from "$lib/kv_primitives.ts";
-import { PrefixedKvPrimitives } from "$lib/prefixed_kv_primitives.ts";
+import { KvPrimitives } from "$lib/data/kv_primitives.ts";
+import { PrefixedKvPrimitives } from "$lib/data/prefixed_kv_primitives.ts";
 import { extendedMarkdownLanguage } from "$common/markdown_parser/parser.ts";
 import { parse } from "$common/markdown_parser/parse_tree.ts";
 import { renderMarkdownToHtml } from "../plugs/markdown/markdown_render.ts";
