+++ server/storage_backend.ts
@@ -2,7 +2,7 @@ import { DiskSpacePrimitives } from "$common/spaces/disk_space_primitives.ts";
 import { SpacePrimitives } from "$common/spaces/space_primitives.ts";
 import { path } from "./deps.ts";
 import { S3SpacePrimitives } from "$common/spaces/s3_space_primitives.ts";
-import { KvPrimitives } from "$lib/kv_primitives.ts";
+import { KvPrimitives } from "$lib/data/kv_primitives.ts";
 import { ChunkedKvStoreSpacePrimitives } from "$common/spaces/chunked_datastore_space_primitives.ts";
 import { HttpSpacePrimitives } from "$common/spaces/http_space_primitives.ts";
