+++ web/client.ts
@@ -42,9 +42,9 @@ import { MainUI } from "./editor_ui.tsx";
 import { cleanPageRef } from "$sb/lib/resolve.ts";
 import { SpacePrimitives } from "$common/spaces/space_primitives.ts";
 import { CodeWidgetButton, FileMeta, PageMeta } from "../type/types.ts";
-import { DataStore } from "$lib/datastore.ts";
-import { IndexedDBKvPrimitives } from "$lib/indexeddb_kv_primitives.ts";
-import { DataStoreMQ } from "$lib/mq.datastore.ts";
+import { DataStore } from "$lib/data/datastore.ts";
+import { IndexedDBKvPrimitives } from "$lib/data/indexeddb_kv_primitives.ts";
+import { DataStoreMQ } from "$lib/data/mq.datastore.ts";
 import { DataStoreSpacePrimitives } from "$common/spaces/datastore_space_primitives.ts";
 import {
   EncryptedSpacePrimitives,
@@ -57,7 +57,7 @@ import {
 import { renderTheTemplate } from "$common/syscalls/template.ts";
 import { PageRef } from "../plug-api/lib/page_ref.ts";
 import { ReadOnlySpacePrimitives } from "$common/spaces/ro_space_primitives.ts";
-import { KvPrimitives } from "$lib/kv_primitives.ts";
+import { KvPrimitives } from "$lib/data/kv_primitives.ts";
 import { builtinFunctions } from "$lib/builtin_query_functions.ts";
 import { ensureAndLoadSettingsAndIndex } from "$common/settings.ts";
 import { LimitedMap } from "$lib/limited_map.ts";
