+++ web/syscalls/clientStore.ts
@@ -1,5 +1,5 @@
 import { SysCallMapping } from "../../lib/plugos/system.ts";
-import { DataStore } from "$lib/datastore.ts";
+import { DataStore } from "$lib/data/datastore.ts";
 import { KvKey } from "$type/types.ts";
 
 export function clientStoreSyscalls(
